WIP to move demo server to its own place
