+++ bokeh-server
@@ -1,10 +1,12 @@
 #!/usr/bin/env python
+
 from __future__ import print_function
-import bokeh.server
-import bokeh.server.start
+
+import bokeh.server.demo
+
 if __name__ == "__main__":
     try:
-        bokeh.server.run()
+        bokeh.server.demo.start()
     except KeyboardInterrupt:
-        bokeh.server.start.stop()
+        bokeh.server.demo.stop()
         print("Shutting down bokeh-server ...")
