+++ bokeh/server/demo/__init__.py
@@ -1,0 +1,3 @@
+
+from .app import start; start
+from .app import stop; stop
