+++ bokeh/server/demo/redis.conf
