+++ bokeh/server/demo/services.py
